+++ ActiveDataProvider.php
@@ -52,6 +52,10 @@ use yii\base\InvalidConfigException;
  * which eliminates ability to verify if requested page number actually exist. Data provider disables `yii\data\Pagination::validatePage`
  * automatically in this case.
  *
+ * Note: because pagination offset and limit may exceed Sphinx 'max_matches' bounds, data provider will set 'max_matches'
+ * option automatically based on those values. However, if [[Query::showMeta]] is set, such adjustment is not performed
+ * as it will break total count calculation, so you'll have to deal with 'max_matches' bounds on your own.
+ *
  * @property array $meta search query meta info in format: name => value.
  * @property array $facets query facet results.
  *
@@ -132,10 +136,20 @@ class ActiveDataProvider extends \yii\data\ActiveDataProvider
         }
         $query = clone $this->query;
         if (($pagination = $this->getPagination()) !== false) {
-            if (!empty($query->showMeta)) {
+            if (empty($query->showMeta)) {
+                $pagination->totalCount = $this->getTotalCount();
+                $limit = $pagination->getLimit();
+                $offset = $pagination->getOffset();
+                // pagination may exceed 'max_matches' boundary producing query error
+                if (!isset($query->options['max_matches'])) {
+                    $query->options['max_matches'] = $offset + $limit;
+                }
+                $query->limit($limit)->offset($offset);
+            } else {
+                // pagination fails to validate page number, if total count is unknown at this stage
                 $pagination->validatePage = false;
+                $query->limit($pagination->getLimit())->offset($pagination->getOffset());
             }
-            $query->limit($pagination->getLimit())->offset($pagination->getOffset());
         }
         if (($sort = $this->getSort()) !== false) {
             $query->addOrderBy($sort->getOrders());
@@ -161,9 +175,11 @@ class ActiveDataProvider extends \yii\data\ActiveDataProvider
             throw new InvalidConfigException('The "query" property must be an instance "' . Query::className() . '" or its subclasses.');
         }
 
-        $meta = $this->getMeta();
-        if (isset($meta['total'])) {
-            return (int) $meta['total'];
+        if (!empty($this->query->showMeta)) {
+            $meta = $this->getMeta();
+            if (isset($meta['total'])) {
+                return (int) $meta['total'];
+            }
         }
 
         $query = clone $this->query;